requirements.txt
Django==5.1.3
dj-database-url
psycopg2-binary
python-dotenv==1.0.1
Pillow==11.1.0
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt
whitenoise